requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu124
numpy
torch
einops
Pillow
timm
torchvision
gradio